fix(tests): do not pin TTL on Tier::Long after #2399 permanence
Check (macos-fed,sqlite) on 9e511414 went RED on two tests that used
Long as a non-default label while also asserting a live expires_at:

- federation_catchup_send_carries_every_durable_field_2384 fixture
  last_accessed_at+expires_at (maximal_memory was Long)
- a5_1579_candidates_path_applies_namespace_and_liveness_filters
  (make_mem is Long; past expires_at is stripped)

#2399 makes a fresh long-tier write permanent. Catch-up fixture now
uses Short (still non-default vs Mid) so expires_at is projected;
the liveness-filter candidate is Mid so TTL is load-bearing.
Production contract unchanged. Never reran 9e511414.

+++ tests/federation_catchup_column_fidelity_2384.rs
@@ -87,8 +87,11 @@ fn fresh_conn() -> rusqlite::Connection {
 fn maximal_memory() -> Memory {
     Memory {
         id: uuid::Uuid::new_v4().to_string(),
-        // Default reconstruction would be `Tier::Mid`.
-        tier: Tier::Long,
+        // Default reconstruction would be `Tier::Mid`. Must NOT be
+        // `Tier::Long`: #2399 makes a fresh long-tier write permanent, so
+        // a caller-supplied `expires_at` is stripped and this fixture
+        // could not pin that column on the catch-up projection.
+        tier: Tier::Short,
         namespace: "team/fed-2384".to_string(),
         title: "federation catch-up column fidelity — #2384".to_string(),
         content: "a Decision-kind claim with a bounded validity window".to_string(),
